requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
sounddevice==0.4.6
scipy==1.12.0
numpy==1.26.4
pyaudio==0.2.14
SpeechRecognition==3.10.1
pydub==0.25.1 